Update patch set 2
Patch Set 2:

Hi Bill,

A review for you if you have the time. Adding actions to the Multicore Visualizer's view menu, I noticed that they were not getting updated like the actions in the context menu. 

In the end, I think all that was needed was to register a listener to have the chance to refresh the actions when the view menu is invoked. I followed closely your context menu implementation - hopefully that makes sense.

Thanks,

Marc

Patch-set: 2