Update patch set 3
Patch Set 3: -Code-Review

Hi,

The previous version of the patch did not work - since the population of the view menu was happening through a listener, when the menu was about to be open, and the view menu is only shown if it contains actions, it ended-up not being shown at all.

I can see two ways to get around this:

1) update both the view menu and toolbar in updateUI(), like it was originally done. We might get double updates in some cases. 
2) do an initial populating of the view menu, when the viewer is set. Then if the viewer has actions to add to the view menu, it will be shown and the listener will be called when the menu is open, updating it / repopulating it as needed. 

This new patch tries to do option #2. 

Comments welcome!

Patch-set: 3