Update patch set 9
Patch Set 9:

(4 comments)

> Do we really have a dependency on: https://git.eclipse.org/r/#/c/36435/18

No, but it's merged now so it doesn't matter anymore :).

> We definitely need to support Windows before moving forward

I will do that.

> It would be nice if this compatibility layer is ready to grow its support to other platforms e.g. OSX, not necessary to proceed with this change as this was not the case with the current implementation.

I don't think there will be a problem to change it to support OS X, for example. Although the "pthread/posix" version is a bit of a lie, because it uses some non-posix functions (for the thread name for example). Therefore, we might need to make "Linux" and "OSX" versions rather than "Pthread". But since most of the things would be the same in both, we could think of ways to make a generic pthread version, and make Linux and OSX inherit of that one.

> The following tests are failing when running with this new structure, however they are fixed on the next commit, Gerrit change 36815.
> This commit shall be squashed with the c/36815 as the commits shall be complete e.g. no compilation or test errors.

Ok, I will squash them.

Patch-set: 9
Label: Verified=0