debug tests: Cleanup filename references
This patch attempts to standardize the naming and factor out the variables
that refer to source or executable files throughout the debug tests.

It removes definitions of paths that are already defined in BaseTestCase.

Also, it tries to name these consistently:

- filename of executable: EXEC_NAME
- filename of source: SOURCE_NAME

Finally, it replaces hardcoded paths at various places by constants at the top
of the test class.

Change-Id: Ib2ea3e46b41185fb9614ae6ad9d41c3b70154884
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>
Reviewed-on: https://git.eclipse.org/r/38068
Reviewed-by: Marc Khouzam <marc.khouzam@ericsson.com>
Tested-by: Hudson CI
Tested-by: Marc Khouzam <marc.khouzam@ericsson.com>