Bug 389536 - NPE in AdditionalInput for projects containing files with custom build-steps.
Tools for custom build steps normally don't belong to a tool chain. We
assume now that the inputs don't need to be rebuilt rather than
provoking a NPE.

The actual fix is similiar to the patch added to the mentioned bug
originally been created by Piotr Aniola. It has been extended with a
test case demonstrating the problem (and the fix).

Change-Id: I5f17248cbe8e84779144bb5f1873c5b9456baa16
Signed-off-by: Sebastian Bauer <mail@sebastianbauer.info>
Reviewed-on: https://git.eclipse.org/r/38206
Tested-by: Hudson CI
Reviewed-by: Elena Laskavaia <elaskavaia.cdt@gmail.com>