Update patch set 6
Patch Set 6:

In fact, the only way 'child' is null if constructor got an error response as MIOutput. In that case, the object is not usable anyway - accessing default values in case of error is quite wrong. It never actually happens, DSF only uses methods of this class if request succeeded, but if anybody fails to check for error in future, I'd argue NPE is better than silently getting wrong data. I can add assert to every method if that's more clear.

Patch-set: 6
Label: Verified=0