debug tests: Disable some reverse tests for gdb 7.0 and 7.1
The following tests fail with gdb 7.0 and 7.1:

  - testStopAtMainWithReverse(Restart)?
  - testStopAtOtherWithReverse(Restart)?

The reason is that execution crosses getenv() while recording is
enabled. gdb has some trouble with that, and outputs an error such as:

	warning: Process record ignores the memory change of instruction at address 0x7ffff7de951f because it can't get the value of the segment register.
	warning: Process record ignores the memory change of instruction at address 0x7ffff7de9576 because it can't get the value of the segment register.
	Process record doesn't support instruction 0xfef at address 0x7ffff7a9e5e2.
	Process record: failed to record execution log.

	[process 6993] #1 stopped.
	0x00007ffff7a9e5e0 in strlen () from /lib/x86_64-linux-gnu/libc.so.6

We could either make the test "easier" to make it pass on those gdb
versions, or disable it for those gdb versions. By "easier", I mean just
execute some simple arithmetic instead of some calls to libc.

I think it is counter-productive to reduce the span of the tests just to
make some old gdb versions happy, so I chose to disable it for those.
Actually, the best would be to write a new test which covers less but
passes for all versions.

Change-Id: I98499fbb5c099232bc39dad3906d7348912b89af
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>