Update patch set 4
Patch Set 4:

I think I've got this down to two of my tests failing, and now I need some further guidance. 

Both problem tests are in IndexCPPBindingResolutionTest, the first is testUserDefinedLiteralBadConcat1 which tests if `auto test = "123"X "123"Y;` correctly fails, in CPreprocessor I set the location of the problem binding to cover the second string literal and with debugging I've checked that it does actually handle the problem. But, in the test, `getProblemFromASTName` (as far as I can tell) only checks if there's a problem on the variable. Is there another assertion I need to make, or is this test only needed in the AST Tests and can be removed from the index biding tests?

The second problem stems from the raw literal operator vs literal operator template resolution in CPPSemantics. I think I may be misunderstanding how to use resolveFunction/resolveAmbiguities, since the candidate set has the exact same function twice, one as a PDOM node and the other as an AST node. How do I remain agnostic about the type of node, yet resolve to the correct one?

Patch-set: 4