Update patch set 3
Patch Set 3:

Discussing this we agree that changing the visibility of those fields from package to private does not break the API. The concern for a possible fragment accessing those fields is a corner case scenario that does not need to be addressed.

So I am ok to merge this improvement as soon as the minor updates are addressed.

Patch-set: 3