Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: I64b408b09444df818d30d99d99de4a1974eacf93
Subject: Bug 399215 - Toggle Function breaks the code... - Test case for reproducing the problem. - The code is lost because a copy of the body, which contains macro references is not rewritten, but its raw signature is taken. The raw signature of a copy is empty in this case. I've solved this issue by using the raw signature of the original node for get raw signature. This is a fundamental change of ASTNode.getRawSignature(). We could also solve this in the StatementWriter to get the original node before accessing the raw signature.
Branch: refs/heads/master
Commit: 8794ac432e8a9cdd612d7bac97833309e0cf1613