Update patch set 2
Patch Set 2:

Very cool!

I think this deserves to be mentioned in the N&N for the user-visible part.

One thing to note is that the package explorer view is part of JDT, and the the Project Explorer that CDT uses does not have a Plugin-in dependencies node.  Is there another way to make use of this nice improvement without installing JDT?

Patch-set: 2