Add setup file for Oomph
This patch adds a setup file for Oomph. To test it:
1. Start the Oomph installer (https://wiki.eclipse.org/Eclipse_Oomph_Installer)
2. At the package selection, click the Advanced Mode button in the top-right
corner
3. Select Eclipse IDE for Eclipse Committers. Click next.
4. Click on the '+' button in the top-right corner, browse for CDT.setup.
5. Double-click on CDT in the tree. Click next.
6. Enter installation folder name, workspace name. Click Next, Finish.

This should install and start a new Eclipse installation with the CDT code
ready to be worked on. Once this patch is merged, the setup will be added to
Eclipse catalog so that the user doesn't have to browse for the file.

Change-Id: Ia84383e9cc64fd3eec0e738e5266dcd8ba6dad0a
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>