Update patch set 4
Patch Set 4:

Thanks for the info.  The build runs with Java 7 so we are ok.  But it is worth noting that the normal jar files are not removed, they are still part of the p2 repo, along with the pack.gz ones.  This is a directive from the SimRel and I believe the reason is actually what you mention.

Patch-set: 4