Create patch set 2
Patch-set: 2
Subject: Bug 399215 - Toggle Function breaks the code... - Test case for reproducing the problem. - The code is lost because a copy of the body, which contains macro references is not rewritten, but its raw signature is taken. The raw signature of a copy is empty in this case. I've solved this issue by using the raw signature of the original node for get raw signature. This is a fundamental change of ASTNode.getRawSignature(). We could also solve this in the StatementWriter to get the original node before accessing the raw signature.
Commit: 786e2137ebbb7679d12ade0ccec08c252400f7cc