Update patch set 2
Patch Set 2:

Regarding Mikhail comment, yeah it is nicer if service would do it, but this is really a -2 (or -2 was for something else)?
I actually checked how we deal set autosolib and this one is just calling command handler directly.
The changes proposed by Mikhail I can add incrementally later,
it just substantially more work and I may not have time for that now. I will address Marc's comments in the code.

Patch-set: 2
Label: Verified=0