Update patch set 7
Patch Set 7:

I am not sure I understand #2. We DO have to run info shared after this command executed because this one does not produce enough data to refresh Modules view. Event would not help any-how in this case. And that is what we need to do anyway. I am not sure what are you proposing?

Patch-set: 7