Update patch set 7
Patch Set 7:

About #2, the problem is that the call to DefaultRefreshAllTarget.refresh() will clear the caches of all the services of any view that has been shown.  For example, if you look at variables, expressions, registers, and then Load Symbols in the Modules view, all 4 views will be affected and all their corresponding data cached in the services will be thrown away.

Maybe doing a refresh just on the Modules view would be ok.  To be honest, I'm a little tired right now and I'm not thinking very clearly.  I'll have to think about this a little more.  Maybe Mikhail can clarify?

Patch-set: 7