Update patch set 6
Patch Set 6:

I may need further guidance to solve that macro problem. If I run it through g++, it issues a warning saying that there needs to be an intervening space between the string and the macro. Should CDT issue a warning too? If so, how would I go about seeing if the macro is defined inside Lexer? 

As it stands, the argument to `f()` in the test receives `IProblemBinding.SEMANTIC_INVALID_TYPE`, which as far as I can tell is correct, but I don't know how to indicate to the user that the suffix has no corresponding operator.

I have some time over the next few days to look into this, although I have the flu at the moment, so I may not be doing my best work.

Patch-set: 6
Label: Verified=0