Update patch set 6
Patch Set 6:

Hmm. Clang and our local variant of g++ 4.9 do not produce any warning. It's possible that the warning is disabled by some global setting. Since g++ still compiles the code even if it issues a warning, CDT has to be able to parse it too.

According to my reading of C++11 standard, "aaa"MACRO should be treated as a preprocessor token, but that would mean that C++11 is not backward compatible with C++98. Possibly we have to keep two alternative token representations '"aaa"MACRO' and '"aaa"', 'MACRO' and chose between them after macro substitution.

I may be able to consult a C++ expert in a couple of days.

Patch-set: 6