Update patch set 6
Patch Set 6:

That was how I understood the spec too. My patch currently assumes c++11 parsing if the scanner has "fSupportsUserDefinedLiterals" set to true. 

If we go the route of having two representations of string/char UDLs, my plan would be to modify `fetchToken` in CPreprocessor to do something like:

if nextToken == tUSER_DEFINED_STRING_LITERAL:
    suffix = StringType.getUserDefinedLiteralSuffix(nextToken)
    type = nextToken.getType()
    if !type and isMacro(suffix):
        nextToken = nextToken.withoutSuffix()
        tokenList.push(suffix)

An alternate way would be introspecting what standard we're using. My understanding of CDT is that it does really "know" what standard it's parsing for. Perhaps this could be changed, and some kind of feature table could be added depending on what __cplusplus is set to. eg, if it's 201104, then CDT enables all c++11 features, if it's lower, then it's set to c++98. If this is a change to be considered, my udl patch may have to wait.

Patch-set: 6