Update patch set 7
Patch Set 7:

Also, can the new 'sharedlibrary' command need to be sent more than once for the same module?  I assume that once a module is loaded, it does not need to be loaded again?  In which case, I'll setup the cache so that 'sharedlibrary <name>' can only be sent once per session for each <name>.

Patch-set: 7