Update patch set 8
Patch Set 8:

> In theory, the Debug view should be refreshed as well
> when symbols for a shared library are loaded. 
> It may contain stack frames without symbol 
> information which should be updated to reflect 
> the changes. Updating the Debug view will force 
> updates of other views.

Using events, we can just update the StackFrameVMNode, which is the one that can change.  That will avoid affecting other views and services.

However, it was mentioned to me that maybe a variable could point to a method for which the symbol is now loaded, so the variables view needs to be refreshed.  Same  could be for Registers, and for Disassembly.

Maybe refreshing all views as Elena's was doing in patchset 7 is not so bad in the end.

Mikhail, what do you think?

Patch-set: 8