Update patch set 10
Patch Set 10:

Patchset 10 puts back the refreshAll call from DsfLoadSymbolsCommandHandler, except that it also refreshes all views when loading all symbols.  I'm not sure why it wasn't doing that before.

An improvement to this will be to go through each VMNode and each service to listen to Symbol*Event and selectively refresh/clear what can change.  This is too big an effort to do today.  Since this seems important for Elena, I'm ok with committing the feature without this improvement if Mikhail is also ok with it (although next time, maybe we can allow for more time for something like this :)).

One thing we must finalize before committing is the new Symbols*Event in IModules2 as they will become API.  Nothing else is API, as far as I can tell.

Patch-set: 10