Update patch set 12
Patch Set 12: Code-Review+1

> Maybe using an empty array to indicate that symbols for all modules are loaded/unloaded. What do you think? Is it confusing?

Funny, I used that exactly if there is an error in getting the list of modules :)  I could use it even in the success case, but it turns out it was very simple to call getModules() to fill the array.

So, pathset 12 is the best I can do for today I think.

I don't like the use refreshAll() as it clears all caches of all services, which is too aggressive, but that can be improved later.

Mikhail?

Elena?  I hope you don't mind I hijacked your code like that.  And I recommend you test that things work as you expect.

Patch-set: 12