Update patch set 1
Patch Set 1:

(10 comments)

This might just work enough to teach users about it.

I'd like to give it a bit more thought to see if we can improve a bit to show actual grouping syntax.

Thanks for doing this.

We'll need to get approval from the list if we are to put this in 8.6

Patch-set: 1