Update patch set 1
Patch Set 1:

Ok I will move it to gdb, I was not sure where it is implemented. I will also would then do visibleWhen checkEnabled, so they don't show up at all when not available.
Putting them in submenu also not a bad idea, but I don't know what happens to menu if non of its items active.

"I am feeling Lucky" is kind of lame. We don't usually put jokes like this in any other part of CDT, so it would be kind of weird. Help is also not a right thing to do in a menu. Eclipse has standard ways to provide help and it should just work. Btw if you want to more visiblity on syntax one place to put it is in Add Watch Expression dialog itself, is it ours or platform?

Patch-set: 1