Update patch set 2
Patch Set 2: Code-Review-1

(5 comments)

Interesting improvement.  It is hard to be clear on what is happening since it is in the error path, which also means I can't really test it.

I'll rely on your results but I put some points on things that may need to be changed.

Patch-set: 2
Label: Code-Review=-1