Bug 458295 - Add SourceReferences to Remote Services plugins
This allows the user to clone and checkout the source code of Remote Services
without knowing where the repository is. See:
https://wiki.eclipse.org/PDE/UI/SourceReferences

One use case is to right-click on a plug-in under the Plug-in dependencies in
Package Explorer view and select Import from Repository. Note that for this
EGit needs to be installed.

What this patch does is to let Tycho generate the source references in the
MANIFEST.MF of the built jars. So to test this patch, one can build the update
site locally and inspect the manifests of the (non-source) jars.

Change-Id: I9214bfc31c6d04a6945be398a2c7b10f9a30d204
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>