Update patch set 7
Patch Set 7:

That test is now passing; one of the recent commits to add "alignof" was using the same IToken values that I was using. 

Anyway, now I'm back to the problem that the macro test fails because of that NegativeArraySizeException I mentioned in the previous patch set. I don't understand why it's saying the token I generate has a length of -87.

Patch-set: 7