Update patch set 8
Patch Set 8:

As far as I can tell, the problems seems to stem from the expandMacro() function, which sets the offset/endOffset to 0/1 of the returned macroToken. This leads me to think that the way I'm expanding the macro is probably the wrong way to do it. 

This macro expansion part is starting to go beyond my depth, to where I'm not sure how to proceed. Is there another way I should be splitting this token up?

Patch-set: 8
Label: Verified=0