Update patch set 1
Patch Set 1:

"1) Do you know if it's possible to configure a the runtime instance to use a specific default perspective e.g. C++ ?, the one used today is Resource"

Always forcing the same perspective is easy (-perspective) but setting the *default* perspective would be harder. The only way I know is with plugin_customization.ini. We would need to add one to the repo and use the -pluginCustomization argument to point to it (using variables?).

"2) When I restart the Eclipse instance I am getting Null pointer exception which does not seem to affect the Eclipse functionality"

I saw this too, I'll report it.

Patch-set: 1