Update patch set 4
Patch Set 4:

I would like to suggests some minor tweaks, which I pushed as a new patchset to make things easy.  You can compare patchset 4 with 3 to see the changes.  If you disagree with any of them, I'll revert them.

One thing I didn't do but I think we should, is to do some sanity on the pid.  I suggest just making sure it is an integer.  We do the same for 'remotePort' in ApplicationWorkbenchWindowAdvisor.java, but I wasn't sure where it was best to do that for the pid.

I also didn't actually test the changes but since you did, that is good enough for me.

Patch-set: 4
Label: Verified=0