Update patch set 4
Patch Set 4:

I didn't test the pid because I found that if it was invalid, the dialog would come up which made more sense to me than failing.  I also thought that the code that was verifying the pid was invalid could have posted a dialog which would have made more sense.

The testing I did was to try -a by itself (which brought up the dialog) and then I did -a pid which brought the debugger up directly for the gedit app I got the pid for.  I also tried an invalid pid and this brought up the dialog.

Patch-set: 4
Label: Verified=0