Bug 458091 - Debug frames - cache
- Implemented cache for stack trace 
(replaces original stack depth cache with a cache of frames and depth)
- This is mostly done to mitigate bugs in gdb 
but it will also help speed up the debugger since the mi cache 
doesn't cache commands with different limits and does not know 
that if we requested frames already we don't need to ask about stack
depth anymore
- In case of unrecoverable errors I will also return some 
constant depth like 5, later the front end will ask to fill it up and in
many cases gdb will be happy to do that, even if it failed to report the proper
depth in the first place. We originally just returned 1 but it is really not
enough for most of the cases. This will result in potential white gaps
at the end of trace, but it is better then showing just the first frame when we have errors

Change-Id: I24b42fd7ffea082e8064a9c5348fd95c5f7777be
Signed-off-by: Alena Laskavaia <elaskavaia.cdt@gmail.com>