Bug 458091 - Debug frames - cache
- Implemented cache for stack trace 
(replaced orginal stack depth cache with cache of frames and depth)
- This mostly done to mitigiate bugs in gdb 
but it will also help speed up debugger since mi cache 
don't cache commands with diffrent limits and does not know 
that if we requested frames already we don't need to ask about stack
depth anymore
- In case of unrecoverage errors I will also return some 
constant depth like 5, later front end will ask to fill it up and in
many cases gdb will be happy to do that, even if fail to report proper
depth in a first place. We orginally just return 1 but it is really not
enough for most of the cases. This will result in potential white gaps
at the end of trace, but it is better then showing just first frame when
we have errors

Change-Id: I24b42fd7ffea082e8064a9c5348fd95c5f7777be
Signed-off-by: Alena Laskavaia <elaskavaia.cdt@gmail.com>