Add build flag (-DskipDoc) to be able to skip building documentation
Building documentation can take several minutes and is often not necessary.
I suggest that we leave it on by default so that people building locally
still get a full, consistent build of CDT but that it should be possible to
skip it on demand.

My build time went from 4:14s to 1:45s (without running tests).

Change-Id: Ic7c4535f3faf82080534d39ccbe6f452d8b52317
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>