Update patch set 1
Patch Set 1: Code-Review+2

Looks good.
Note that because we are porting this change to 8.6, we'll need to update the version of this plugin on master to 1.2.  But we can wait until after the 8.6 release to properly set the API baseline and the API Tool will tell us how to fix it properly.

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=+2