[RFC] Support display as array with dynamic lengths
The goal is to be able to specify a variable or an expression as the
length of the array when using "Display as array". This can be useful
when debugging a C program where an array is passed to a function with
the length as an argument beside it.

It doesn't work completely right now. It works to specify an expression,
and the initial length corresponds to the result of the expression, but
the length is not updated when the result of the expression. The
interface keeps displaying the initial number of elements.

Change-Id: I0ca5e25a74807aed4ffd69c95a23e86f1460d86d
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>