Allow for filtering of findbugs
Choose version 2.5.2 which is what is supported by CDT's HIPP.  Note that by removing the version altogether, the latest available findBugs would be selected.  However, this could lead to different results between manual builds and HIPP builds, which is why I thought it would be better to specify the version explicitly.

FindBugs can be run by adding the string

 findbugs:findbugs

as a maven parameter, either in a local maven build, or on HIPP.

Add releng/findbugs-exclude.xml to control the errors shown by FindBugs.
The content of that file was copied from the TM project, but should
eventually be ajusted for CDT.

Change-Id: Ibc8348b9286f00580a4079a7ca3c669708bc6194
Signed-off-by: Marc Khouzam <marc.khouzam@ericsson.com>