Update patch set 2
Patch Set 2: Code-Review-1

(8 comments)

I'm not against this but it seems very specific to your current need.  I wonder if we can use a more generic way.

For example, the raw output you need is available through MIInfo.getMIOutput() which could be accessed when getting the reply to MIVarCreate, MIVarListChildren, and MIVarUpdate.

I cannot tell if this is possible for the extending you are trying to do, but maybe you can have a look?

I also put in-line comments if we go with your original solution.

Patch-set: 2
Label: Code-Review=-1