Bug 459971 New SerialPort support. First step is Mac library.
Java classes are defined and a pretty simple native library for
Mac is there. Tested with Arduino so I know reading works. This
implementation should work for Linux as well. Windows is going to
be the hard one.

Not checking in the binaries yet until I get more testing with them.

Also remove ppc native libraries for Mac since that hasn't been
supported in a long time.

Change-Id: If4ffbc6e73a7656a47c2f45b875be0842c482b05