Bug 293832 - [dsf-gdb] Ask gdb for variable value, even if it is a complex type
Currently, CDT does not ask GDB for the value of the variables it
believes to be of complex types, such as structures. However, as
described in the bug, it assumes that a typedefed pointer is a complex
structure. Because of that, it displays a value of "{...}" for it
instead of the pointer value.

By asking GDB for the value of the variable even if it's of a complex
type, CDT will always display the right thing for the value. This will
cause a few more -var-evaluate-expression calls, but their number is
still limited to what is visible in the variables view. So the impact
should be negligible.

Screenshot of before/after:
	http://i.imgur.com/Sx5ZPfO.png

It is still impossible to edit the value, but this would be addressed in
another patch.

Change-Id: I92e6ead6351677e098a56d4af5bdb85fc61df080
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>