Update patch set 3
Patch Set 3:

I would recommend not passing around an MIVariableObject.  Although the type is public (not sure why), variable objects are an implementation detail and should not spill over MIVariableManager.

We can argue that ExprMetaGetVarInfo is so closely related to Variable Objects that the spill over is justified, so I won't oppose it, but my instincts tell me that we should keep that type inside MIVariableManager as much as we can.

How bad is it to not pass it to ExprMetaGetVarInfo?

Patch-set: 3