Update patch set 6
Patch Set 6: Code-Review+2 Verified+1

> Does this mean we can run findbugs plugin on hudson? Or this
 > is just dormant code?

Yes, we can run it on Hudson by adding findbugs:findbugs to the maven command-line parameter.  I was thinking of enabling it for the nightly jobs (master mostly).

Doing that slows down the build by about 20 minutes so I was trying to get the master build to run findbugs when triggered automatically at night, but prompt when the job is triggered manually.  The idea is that in some cases we need a successful build but don't care to wait for FindBugs to run.

I was playing around to get that to work but I haven't found a good way yet.

This patch is independent of that though and could go in now.

Patch-set: 6
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=+1
Label: Code-Review=+2