Bug 382746 - Adding C/C++ nature for C Project impossible
This patch modifies the wizard so that the C projects also show in the
list of projects.
- If a given C project is selected and the C nature is selected, the
project is not modified.
- If a given C project is selected and the C++ nature is selected, the
C++ nature is added and project type and toolchain selection have no
effect because it is assumed that the project type and toolchain had
already been setup before hand.
Unchanged behavior:
- C++ projects never show in the list of projects. Removing nature is
(still) unsuported.
- Non-CDT project behave as they use to.

Change-Id: Ie6282c11d90c42d21ecad2996ab49ebd64c38ece
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>