Update patch set 3
Patch Set 3:

@Elena, Upon re-reading, I think I got my messages mixed up. 

I suggested removing the sleep in ps1. 

You replied: swtbot waituntil may run forever

I suggested if timeouts are a concern, having a global timeout for the test.

You replied that sleeps are a bad practice. (which I am 100% in agreement with) 

I would posit though and probably not for this patch, that if it is not done in some hidden code somewhere already, having a timeout for tests with a good factor of safety, would allow hudson (or jenkins) to not hang on a build. I think though we are discussing different items, I agree, waitUntil(condition, timeout) is the way to go.

SWTBotPreferences.Timoute is the default value of the timeout though.

<code>
public void waitUntil(ICondition condition) throws TimeoutException {
	waitUntil(condition, SWTBotPreferences.TIMEOUT);
}
</code>

Patch-set: 3