Update patch set 1
Patch Set 1:

At least one of them failed when testing with an architecture of fWordSize = 2. The test expected another error to be thrown, but the service instead complained that the input buffer was not big enough (probably because that condition is checked before in the code). So, unless we want to test that specifically (which is the point of writeWithInvalidBuffer), I thought it was better to remove the ambiguity on which error is thrown.

Patch-set: 1