Update patch set 3
Patch Set 3:

(8 comments)

Elegant change.  I have some minor in-line comments.
I'm also gonna try to play around with the generics to see if we can simplify, but not right now, it's getting late.  I'll won't spend too much time on it though, so as not to delay you.

Patch-set: 3
Label: Verified=0