Bug 459971 Windows native for Serial Port.
Not checking the binaries in until I get closer to the end. Still some
API changes I want to make to do more buffering.

Done as a Visual Studio 2013 project. I also brough winreg into that
sln file as I though I had to make a change in it but didn't in the
end. But really should bring them all.

Change-Id: I6e7d472763381cdc0ae558d8cd63993bb0460457
Signed-off-by: Doug Schaefer <dschaefer@qnx.com>