Create patch set 2
Uploaded patch set 2.

Patch-set: 2
Subject: Bug 461628 Some tools within a toolchain can potentially generate an output resource with the same name even if the input resource name to the tool is not the same. This fix is to provide this capability to CDT
Commit: b699723742841b95b7be64a7639ea1f0836df499