Update patch set 1
Patch Set 1:

I do not have principal objections against it, just a grudge that we should not keep adding anybody's favorite tool in the list. For example, somebody wanted to add ccache, see https://bugs.eclipse.org/bugs/show_bug.cgi?id=378984.
It really depends on judgement - if a tool is used by most of developers we should add it. If it is used by just a few - we should not. I use libtool or ccache very rarely so naturally I am suspicious to the requests. But if you, Marc-Andre, and other committers think it is worthy, go ahead.

Patch-set: 1