Update patch set 1
Patch Set 1: Code-Review-1

Briefly looking, it seems that this is the right way to do it. Except that this change caused test failures - which should be easy to fix. Also, please make sure that the corner case when the last line of output ends with backslash is addressed.

Patch-set: 1
Label: Code-Review=-1