Update patch set 1
Patch Set 1:

No, I think we should go with this solution which is why I approved it.  We would still have the issue that tools are called within the make file and could conceivably react to the V=1 environment variable themselves.  Just venting at the lack of forethought in the Autotools maintenance group.

Patch-set: 1