Update patch set 5
Patch Set 5:

(2 comments)

Sorry,

Indeed it might seems strange that I put back the original patch. But as both you and Marc-Andre commented the code is dead as it is now. However I do use the API in the ordering implementation in another fix. For now I agree. Let remove the toOSString.

Patch-set: 5
Label: Verified=0