Update patch set 5
Patch Set 5:

I agree with you and this is in fact the right way to do it. I haven’t looked at the caller since more than a year and I really don’t remember why but I had to call the API with the OS string separator definition. It might be that I introduced the wrong way to do the call while patching the Managed Builder. I’ll have a look at the code again when I will commit the modification to support the ordering of the inputs and make the appropriate modifications to support an agnostic path definition.

Patch-set: 5