Update patch set 6
Patch Set 6:

In the step into selection tests, the test has a .cpp and a .h. We want to have tags in both files, so we need to call it for both files.

Lines will not collide, because the mapping is

  tag -> line

so it's ok to have

  some_tag_in_the_cpp -> 5
  some_tag_in_the_h -> 5

As long as there is no tag being repeated, it's fine. And even then, we could include the filename in the key:

  (file.cpp, some_tag) -> 3
  (file.cpp, some_tag) -> 4

Does it make more sense?

Patch-set: 6
Reviewer: Gerrit User 1069 <1069@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0