Update patch set 1
Patch Set 1:

Do we want to have two separate register-handling services (and advertise both) or is it an implementation detail?  If it is about design and modularity, maybe we can look into using the same pattern as for the GDBPatternMatchingExpressions service?

The idea would be to have a single advertised service to handle registers (GDBManagedRegisterGroups) which would then delegate to MIRegisters (or maybe even GDBRegisters) whenever necessary.  The pattern is similar to overriding MIRegisters service, except it de-couples things in such a way that you can override both MIRegisters (and even GDBManagedRegisterGroups) while keeping GDBManagedRegisterGroups as the first entry point.

I don't know if that patterns fits well in the goals of this patch, but it would make the publishing of the register services a little cleaner.

Patch-set: 1
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0