Bug 463844 - Support serial port connection type.
This adds a remote connection type that uses the CDT Native serial
port driver. Really all this is supported is the command shell for
the Remote Terminal. But it's a great example of something very
different from SSH and Local that supports the same common infra
structure such as the Terminal.

Change-Id: I813e26370d3d137d5158fa63deba4b0820cdd596
Signed-off-by: Doug Schaefer <dschaefer@qnx.com>